Moved getStateForUser_andExpectStateToBeCreated to CurrentPlayerStateTest
